html/semantics/forms/the-input-element/password.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Password input element</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/forms.html#password-state-%28type=password%29">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input id="password" type="password" />
<input id="password_with_value" type="password" value="foobar" />
<script type="text/javascript">
    setup(function() {
      var password = password;
    });

    test(function() {
      assert_equals(password.value, "");
      assert_equals(document.getElementById("password_with_value").value, "foobar");
    }, "Value returns the current value for password");

    test(function() {
      password.value = "A";
      assert_equals(password.value, "A");
      password.value = "B";
      assert_equals(password.value, "B");
    }, "Setting value changes the current value for password");

    test(function() {
      // Any LF (\n) must be stripped.
      password.value = "\nAB";
      assert_equals(password.value, "AB");
      password.value = "A\nB";
      assert_equals(password.value, "AB");
      password.value = "AB\n";
      assert_equals(password.value, "AB");

      // Any CR (\r) must be stripped.
      password.value = "\rAB";
      assert_equals(password.value, "AB");
      password.value = "A\rB";
      assert_equals(password.value, "AB");
      password.value = "AB\r";
      assert_equals(password.value, "AB");

      // Any combinations of LF CR must be stripped.
      password.value = "\r\nAB";
      assert_equals(password.value, "AB");
      password.value = "A\r\nB";
      assert_equals(password.value, "AB");
      password.value = "AB\r\n";
      assert_equals(password.value, "AB");
      password.value = "\r\nA\n\rB\r\n";
      assert_equals(password.value, "AB");
    }, "Value sanitization algorithm should strip line breaks for password");
</script>
